src/requirements.txt
Django==2.0
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
xmltodict==0.11.0
pyEchosign==1.0.1
WeasyPrint==0.42.2
